ai_backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Image processing
Pillow>=10.0.0
shapely>=2.0.0

# Hugging Face dependencies
transformers>=4.30.0
huggingface-hub>=0.16.0
torch>=2.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
